Move Hash to core
